@@ -1,11 +1,11 @@
 {
-  "name": "freshland",
+  "name": "client-freshland-se",
   "version": "0.1.0",
   "private": true,
   "scripts": {
     "dev": "next dev -p 8080",
     "build": "next build",
-    "start": "next start",
+    "start": "next start -p 8080",
     "lint": "next lint"
   },
   "dependencies": {
@@ -13,7 +13,7 @@
     "cookie": "^0.6.0",
     "jquery": "^3.7.1",
     "js-cookie": "^3.0.5",
-    "next": "14.1.3",
+    "next": "14.2.5",
     "react": "^18",
     "react-calendar": "^5.0.0",
     "react-dom": "^18",
@@ -37,9 +37,9 @@
     "@types/react-dom": "^18",
     "@types/react-slick": "^0.23.13",
     "eslint": "^8",
-    "eslint-config-next": "14.1.3",
+    "eslint-config-next": "14.2.5",
     "react-slick": "^0.30.2",
     "stripe": "^15.4.0",
-    "typescript": "^5"
+    "vanilla-cookieconsent": "^3.0.0"
   }
 }
