@@ -8,7 +8,7 @@ import { addToCart, updateCartQuantity, removeCartItem, getSubscriptionOptions }
 import LazyLoad from "react-lazyload"; // Import LazyLoad component
 import { productdetailTranslation } from '@/locales';
 import Link from "next/link";
-import { useRouter } from "next/router"; 
+import { useRouter } from 'next/navigation';
 import Loader from "@/components/atoms/loader/loader";
 
 const lang = process.env.NEXT_PUBLIC_LANG || 'dk';
@@ -42,7 +42,7 @@ const Description = ({ productDetailProps }) => {
       quantity.toString(),
       isOneTimePurchaseActive ? "" : deliveryOption,
     );
-    setLoading(false); 
+    setLoading(false);
   };
   const [deliveryOption, setDeliveryOption] = useState("");
 
@@ -83,6 +83,9 @@ const Description = ({ productDetailProps }) => {
     getSubOptions();
   }, [productId]);
 
+  const categories = productDetailProps?.productDetail?.categories ?? [];
+  const joinedCategories = categories.join(',');
+
   return (
     <>
       <div className={styles.containerBody}>
@@ -113,7 +116,7 @@ const Description = ({ productDetailProps }) => {
                   height={78}
                 />
               </div>
-              <div className={styles.mainImage}>
+              {productDetail && <div className={styles.mainImage}>
                 <Image
                   src={productDetail.thumbnail}
                   alt="Main Image"
@@ -122,11 +125,12 @@ const Description = ({ productDetailProps }) => {
                   priority
                 />
               </div>
+              }
             </div>
             <div className={styles.rightSide}>
               <div className={`${styles.productTitle}`}>
                 <div>
-                  <h3>{productDetail.name}</h3>
+                  {productDetail && <h3>{productDetail.name}</h3>}
                   {/* <Image
                   src="/mockImage/Spain Flag.png"
                   width={40}
@@ -134,31 +138,10 @@ const Description = ({ productDetailProps }) => {
                   alt="Spain Flag"
                 /> */}
                 </div>
-                <h5 dangerouslySetInnerHTML={{ __html: productDetail.price }} />
+                {productDetail && <h5 dangerouslySetInnerHTML={{ __html: productDetail.price }} />}
               </div>
-              <div className={styles.featuresWrapper}>
-                <div className={styles.textWrapper}>
-                  <TextWrapper text="A whole 8 kg of organic oranges from Andalucia 🍊" />
-                </div>
-                <div className={`${styles.textWrapper}`}>
-                  <TextWrapper text="Store cool at 4-8°C 🌡️" />
-                </div>
-                <div className={styles.textWrapper}>
-                  <TextWrapper text="Note: The lemons are class 2, so surface scratches may occur" />
-                </div>
-                <div className={`${styles.textWrapper}`}>
-                  <TextWrapper text="Lasts up to 4 weeks 🏡" />
-                </div>
-                {/* <div className={styles.rightButtonWrapper}>
-                  </div> */}
-              </div>
-              <div className={styles.descriptionWrapper}>
-                <span>{pdt.freeShippingOver}</span>
-
-                <span>
-                  {pdt.OrderBefore}
-                </span>
-              </div>
+              {productDetail?.excerpt && <div dangerouslySetInnerHTML={{ __html: productDetail?.excerpt }}>
+              </div>}
               <div className={styles.boxWrapper}>
 
                 {subscriptionOpt && subscriptionOpt.is_subscription && subscriptionOpt.subscription_schemes && subscriptionOpt.subscription_schemes.length > 0 && (
@@ -243,18 +226,18 @@ const Description = ({ productDetailProps }) => {
                     </button>
                   </div>
                   <div className={`M-Body-Large ${styles.basketWrapper}`}>
-                 
-                      <span
-                        onClick={handleAddToCart}
-                      >
-                        {loading ? <Loader /> : null}
-                        {isOneTimePurchaseActive ? "Add To Basket" : "Subscribe"}
-                      </span>
+
+                    <span
+                      onClick={handleAddToCart}
+                    >
+                      {loading ? <Loader /> : null}
+                      {isOneTimePurchaseActive ? "Add To Basket" : "Subscribe"}
+                    </span>
                   </div>
                 </div>
                 <div className={styles.itemCatagory}>
                   <p>
-                    {pdt.itemNumberSKU} F176 <br />{pdt.category} FRUIT
+                    {pdt.itemNumberSKU} F176 <br />{pdt.category} {joinedCategories}
                   </p>
                 </div>
               </div>
@@ -281,35 +264,35 @@ const Description = ({ productDetailProps }) => {
                     </h3>
                   </div>
                   <span>
-                  {pdt.theseOrange}
+                    {pdt.theseOrange}
                   </span>
 
                   <br />
                   <span>
-                 {pdt.orangeContent}
+                    {pdt.orangeContent}
                   </span>
                 </div>
 
                 <div className={styles.storageWrapper}>
                   <h4>{pdt.storage}</h4>
                   <span>
-                 {pdt.orangeTemp}
+                    {pdt.orangeTemp}
                     <br />
-                   {pdt.byStoring}
+                    {pdt.byStoring}
                   </span>
                 </div>
 
                 <div className={styles.storageWrapper}>
                   <h4>{pdt.ourEconomy}</h4>
                   <span>
-                {pdt.oneOfour}
+                    {pdt.oneOfour}
                   </span>
                 </div>
 
                 <div className={styles.storageWrapper}>
                   <h4>{pdt.season}</h4>
                   <span>
-                  {pdt.orangeThrive}
+                    {pdt.orangeThrive}
                   </span>
                 </div>
                 <div className={styles.nutriWrapper}>
