@@ -11,7 +11,7 @@ const toastTimer = parseInt(process.env.NEXT_PUBLIC_TOAST_TIMER);
 
 const lang = process.env.NEXT_PUBLIC_LANG || 'dk';
 
-function Orderaccount({ showOrderView, setShowOrderView, orders, orderobj, isUserLoggedIn }) {
+function Orderaccount({ showOrderView, setShowOrderView, orders, orderobj, orderDates, isUserLoggedIn }) {
   const mat = myaccountTranslation[lang];
   const [showRequest, setShowRequest] = useState();
   const handleRequest = () => {
@@ -66,6 +66,8 @@ function Orderaccount({ showOrderView, setShowOrderView, orders, orderobj, isUse
     return number;
   }
 
+  const { delivery_dates: deliveryDates, order_date: orderDate } = orderDates || {};
+
   return (
     <>
       {!showOrderView && (
@@ -84,50 +86,42 @@ function Orderaccount({ showOrderView, setShowOrderView, orders, orderobj, isUse
               </thead>
               <tbody>
                 {
-                  orders && orders.length > 0 && orders.map((od, i) => {
-                    return (
-                      <tr key={i}>
-                        <td onClick={() => { handleViewClick(od.order_number) }} style={{ cursor: 'pointer' }}>#{od.order_number}</td>
-                        <td>{od.order_date}</td>
-                        <td>{od.delivery_date}</td>
-                        <td>{od.order_status}</td>
-                        <td>{od.order_total}</td>
-                        <td className={styles.btnlink}>
-                          <Link
-                            href="#"
-                            className={styles.view}
-                            onClick={() => { handleViewClick(od.order_number) }}
-                          >
-                            {mat.view}
-                          </Link>
-                          <Link href="#" className={styles.pdf} onClick={() => {
-                            downloadInvoicePdf(od.order_number);
-                          }}>
-                            {mat.pDFInvoice}
-                          </Link>
-                          <Link href="#" className={styles.request}>
-                            {mat.requestaComplaint}
-                          </Link>
-                        </td>
-                      </tr>
-                    )
-                  })
+                  orders && orders.length > 0 && orders.map((od, i) => (
+                    <tr key={i}>
+                      <td onClick={() => handleViewClick(od.order_number)} style={{ cursor: 'pointer' }}>#{od.order_number}</td>
+                      <td>{od.order_date}</td>
+                      <td>{od.delivery_date}</td>
+                      <td>{od.order_status}</td>
+                      <td>{od.order_total}</td>
+                      <td className={styles.btnlink}>
+                        <Link
+                          href="#"
+                          className={styles.view}
+                          onClick={() => handleViewClick(od.order_number)}
+                        >
+                          {mat.view}
+                        </Link>
+                        <Link href="#" className={styles.pdf} onClick={() => downloadInvoicePdf(od.order_number)}>
+                          {mat.pDFInvoice}
+                        </Link>
+                        <Link href="#" className={styles.request}>
+                          {mat.requestaComplaint}
+                        </Link>
+                      </td>
+                    </tr>
+                  ))
                 }
               </tbody>
             </table>
           </div>
-
-
-
         </section>
       )}
 
       {showOrderView && orderobj && (
-        <section className={styles.orderviewpage} style={{ display: showRequest ? 'none' : 'Block' }}>
+        <section className={styles.orderviewpage} style={{ display: showRequest ? 'none' : 'block' }}>
           <div className={styles.ordercomplete}>
             <p className="W-body-Large">
-              Order #{orderobj.order_data.order_number} was carried out {orderobj.order_data.order_date} and is currently
-              {orderobj.order_data.order_status}.
+              Order #{orderobj.id} was carried out on {new Date(orderobj.date_created).toLocaleDateString()} and is currently {orderobj.status}.
             </p>
           </div>
           <div className={styles.orderviewdetail}>
@@ -135,30 +129,35 @@ function Orderaccount({ showOrderView, setShowOrderView, orders, orderobj, isUse
               <h5>{mat.orderInfo}</h5>
               <div className={styles.orderDetail}>
                 <p className={styles.label}>{mat.orderLabel}</p>
-                <p className={styles.id}>#{orderobj.order_data.number}</p>
+                <p className={styles.id}>#{orderobj.id}</p>
               </div>
               <div className={styles.orderDetail}>
                 <p className={styles.label}>{mat.date}</p>
-                <p className={styles.value}>{new Date(orderobj.order_data.date_created.date).toLocaleDateString()}</p>
+                <p className={styles.value}>{orderDate}</p>
               </div>
               <div className={styles.orderDetail}>
                 <p className={styles.label}>{mat.delivery}</p>
-                <p className={styles.value}>{new Date(orderobj.order_data.meta_data.find(meta => meta.key === 'delivery_date').value).toLocaleDateString()}</p>
+                <p className={styles.value}>{deliveryDates && Object.keys(deliveryDates)[0]}</p>
               </div>
               <div className={styles.orderDetail}>
                 <p className={styles.label}>{mat.status}</p>
-                <p className={styles.value}>{orderobj.order_data.status}</p>
+                <p className={styles.value}>{orderobj.status}</p>
               </div>
               <div className={styles.orderDetail}>
                 <p className={styles.label}>{mat.total}</p>
                 <p className={styles.ordertotal}>
                   {orderobj.totals.currency_code}
-                  <span dangerouslySetInnerHTML={{ __html: orderobj.totals.total_price }}></span>
-                  units
+                  <span> {getCorrectPrice(orderobj.totals.total_price, orderobj.totals.currency_minor_unit)}</span>
                 </p>
               </div>
               <div className={styles.orderDetail}>
                 <p className={styles.label}>{mat.actions}</p>
+              </div>
+              <div className={styles.orderbtn}>
+                <button type="button" onClick={() => downloadInvoicePdf(orderobj.id)}>{mat.pDFInvoice}</button>
+                <button type="submit" className={styles.requestbtn} onClick={handleRequest}>
+                  {mat.requestaComplaint}
+                </button>
               </div>
             </div>
 
@@ -173,32 +172,26 @@ function Orderaccount({ showOrderView, setShowOrderView, orders, orderobj, isUse
                     </tr>
                   </thead>
                   <tbody>
-                    {orderobj &&
-                      orderobj?.items &&
-                      orderobj?.items.length > 0 &&
-                      orderobj.items.map(item => (
-                        <tr className={styles.topRow} key={item.key}>
-                          <td>
-                            {item.name} <span>x {item.quantity}</span>
-                          </td>
-                          <td>
-                            {orderobj.totals.currency_code}
-                            <span>{getCorrectPrice(parseInt(item.totals.line_total) + parseInt(item.totals.line_total_tax), item.totals.currency_minor_unit)}</span> / <span>week</span>
-                          </td>
-                        </tr>
-                      ))}
+                    {orderobj.items.map(item => (
+                      <tr className={styles.topRow} key={item.id}>
+                        <td>
+                          {item.name} <span>x {item.quantity}</span>
+                        </td>
+                        <td>
+                          {orderobj.totals.currency_code} {getCorrectPrice(parseInt(item.totals.line_total) + parseInt(item.totals.line_total_tax), item.totals.currency_minor_unit)}
+                        </td>
+                      </tr>
+                    ))}
                     <tr>
                       <td>{mat.subTotal}</td>
                       <td>
-                        {orderobj.totals.currency_code}
-                        <span>{getCorrectPrice(parseInt(orderobj.totals.total_items) + parseInt(orderobj.totals.total_tax), orderobj.totals.currency_minor_unit)}</span>
+                        {orderobj.totals.currency_code} {getCorrectPrice(orderobj.totals.total_items, orderobj.totals.currency_minor_unit)}
                       </td>
                     </tr>
                     <tr>
                       <td>{mat.shipment}</td>
                       <td>
-                        {orderobj.totals.currency_code}
-                        <span>{getCorrectPrice(parseInt(orderobj.totals.total_shipping) + parseInt(orderobj.totals.total_shipping_tax), orderobj.totals.currency_minor_unit)}</span>
+                        {orderobj.totals.currency_code} {getCorrectPrice(orderobj.totals.total_shipping, orderobj.totals.currency_minor_unit)}
                       </td>
                     </tr>
                     <tr>
@@ -206,130 +199,54 @@ function Orderaccount({ showOrderView, setShowOrderView, orders, orderobj, isUse
                         <label>{mat.total}</label>
                       </td>
                       <td>
-                        {orderobj.totals.currency_code}
-                        <span>{orderobj.totals.total_price}</span>
+                        {orderobj.totals.currency_code} {getCorrectPrice(orderobj.totals.total_price, orderobj.totals.currency_minor_unit)}
                       </td>
                     </tr>
                   </tbody>
                 </table>
               </div>
             </aside>
-
-          </div>
-          <div className={styles.orderbtn}>
-            <button type="button" onClick={() => {
-              downloadInvoicePdf(orderobj.order_number);
-            }}>{mat.pDFInvoice}</button>
-            <button type="submit" className={styles.requestbtn} onClick={handleRequest}>
-              {mat.requestaComplaint}
-            </button>
           </div>
 
-          {/* detail */}
           <div className={styles.viewtabledetail}>
-            {/* <table className={styles.viewtable}>
-              <thead>
-                <tr>
-                  <th>{mat.subs}</th>
-                  <th>{mat.status}</th>
-                  <th>{mat.nextpayment}</th>
-                  <th>{mat.total}</th>
-
-                </tr>
-              </thead>
-              <tbody>
-                <tr>
-                  <td>#{orderobj.order_number}</td>
-                  <td>{orderobj.order_date}</td>
-                  <td></td>
-                  <td>{orderobj.order_status}</td>
-                  <td className={styles.btnlink}>
-                    <Link
-                      href="#"
-                      className={styles.view}
-                    // onClick={handleViewClick}
-                    >
-                      {mat.view}
-                    </Link>
-                  </td>
-                </tr>
-
-
-
-
-
-              </tbody>
-            </table> */}
             <div className={styles.addresscontent}>
               <div>
                 <p className={`W-Body-Regular ${styles.followingpage}`}>
                   {mat.theFollowingAddresses}
                 </p>
               </div>
-
               <div className={styles.addresscontainer}>
                 <div>
                   <h4>{mat.billingAddress}</h4>
                   <div>
-                    {(orderobj?.order_data?.billing || orderobj?.order_data?.shipping) && (
-                      <AddressInfo address={orderobj?.order_data?.billing || orderobj?.order_data?.shipping} />
+                    {orderobj.billing_address && (
+                      <AddressInfo address={orderobj.billing_address} />
                     )}
                   </div>
                 </div>
                 <div>
                   <h4>{mat.deliveryAddress}</h4>
                   <div>
-                    {orderobj?.order_data?.shipping && <AddressInfo address={orderobj?.order_data?.shipping} />}
+                    {orderobj.shipping_address && <AddressInfo address={orderobj.shipping_address} />}
                   </div>
                 </div>
               </div>
             </div>
-
           </div>
-
-
         </section>
       )}
 
-      {/* for mobile view */}
       {!showOrderView && orderobj && (
-        <div className={styles.mobileviewtable} onClick={handleViewClick}>
+        <div className={styles.mobileviewtable} onClick={() => handleViewClick(orderobj.id)}>
           <div className={styles.tablecard}>
             <div className={styles.ordercarddetail}>
               <div className={styles.orderid}>
-                <p>{mat.orderNo} <span className={styles.status}>{orderobj.order_data.number}</span></p>
-                <p className={styles.status}>{orderobj.order_data.status}</p>
+                <p>{mat.orderNo} <span className={styles.status}>{orderobj.id}</span></p>
+                <p className={styles.status}>{orderobj.status}</p>
               </div>
               <div className={styles.date}>
-                <p>{new Date(orderobj.order_data.date_created.date).toLocaleDateString()}</p>
-                <p>{orderobj.totals.currency_symbol} for
-                  <span dangerouslySetInnerHTML={{ __html: orderobj.totals.total_price }}></span> units
-                </p>
-              </div>
- 
-            </div>
-            <div className={styles.ordercarddetail}>
-              <div className={styles.orderid}>
-                <p>{mat.orderNo} <span className={styles.status}>{orderobj.order_data.number}</span></p>
-                <p className={styles.status}>{orderobj.order_data.status}</p>
-              </div>
-              <div className={styles.date}>
-                <p>{new Date(orderobj.order_data.meta_data.find(meta => meta.key === 'delivery_date').value).toLocaleDateString()}</p>
-                <p>{orderobj.totals.currency_symbol} for
-                  <span dangerouslySetInnerHTML={{ __html: orderobj.totals.total_price }}></span> units
-                </p>
-              </div>
-            </div>
-            <div className={styles.ordercarddetail}>
-              <div className={styles.orderid}>
-                <p>{mat.orderNo} <span className={styles.status}>{orderobj.order_data.number}</span></p>
-                <p className={styles.status}>{orderobj.order_data.status}</p>
-              </div>
-              <div className={styles.date}>
-                <p>{new Date(orderobj.order_data.date_created.date).toLocaleDateString()}</p>
-                <p>{orderobj.totals.currency_symbol} for
-                  <span dangerouslySetInnerHTML={{ __html: orderobj.totals.total_price }}></span> units
-                </p>
+                <p>{new Date(orderobj.date_created).toLocaleDateString()}</p>
+                <p>{orderobj.totals.currency_symbol} {getCorrectPrice(orderobj.totals.total_price, orderobj.totals.currency_minor_unit)}</p>
               </div>
             </div>
           </div>
@@ -337,11 +254,7 @@ function Orderaccount({ showOrderView, setShowOrderView, orders, orderobj, isUse
       )}
       <button style={{ display: showRequest ? 'block' : 'none' }} onClick={handleGoBack} className={styles.gobackbtn}><i className="fa-solid fa-circle-arrow-left"></i></button>
 
-      {
-        showRequest &&
-        <OrderProcess />
-      }
-
+      {showRequest && <OrderProcess />}
     </>
   );
 }
