@@ -58,9 +58,11 @@ const Register = () => {
       <div className={registerStyles.gridContainer}>
         <div className={registerStyles.registerWrapper}>
           <Image
-            src="/Images/regitserimg.png"
+            src="/Images/rimage.png"
             alt="rimage"
-           fill
+            width={573}
+            height={811}
+            className={registerStyles.imgWrapper}
           />
           <div className={registerStyles.innerWrapper}>
             <h2 className={`W-H2 ${registerStyles.heading}`}>{hpt.subscribeNewsletter}</h2>
@@ -104,9 +106,7 @@ const Register = () => {
                 <button className={registerStyles.registerButton} type="submit" onClick={registerform}>
                   {hpt.register}
                 </button>
-               
-              </div>
-              <div className={registerStyles.checkboxContainer}>
+                <div className={registerStyles.checkboxContainer}>
                   <input
                     type="checkbox"
                     className={registerStyles.checkbox}
@@ -116,10 +116,11 @@ const Register = () => {
                     {hpt.iAgree}
                   </label>
                 </div>
+              </div>
             </form>
           </div>
         </div>
-        {/* <div className={registerStyles.gridItem}>
+        <div className={registerStyles.gridItem}>
           <div className={registerStyles.contactWrapper}>
             <h1 className={`W-H3 ${registerStyles.contactHead}`}>{hpt.contactUs}</h1>
             <ul className={registerStyles.contactList}>
@@ -142,7 +143,7 @@ const Register = () => {
               </li>
             </ul>
           </div>
-        </div> */}
+        </div>
       </div>
     </div>
   );
