@@ -3,8 +3,9 @@ import Header from "@/components/atoms/Header/Header";
 import styles from "./Order.module.css";
 import { orderconfirmationTranslation } from '@/locales';
 import Link from "next/link";
-import { getCheckoutOrderById } from "@/components/service/account"
+import { getCheckoutOrderById, getOrderDates } from "@/components/service/account"
 import { AddressInfo } from "@/components/atoms/address/address";
+import Orderskeleton from "@/components/skeleton/orderskeleton";
 
 const lang = process.env.NEXT_PUBLIC_LANG || 'dk';
 
@@ -15,20 +16,28 @@ const Order = ({ orderId }) => {
     cartTotalDiscount: 0,
     shippingTotal: 0,
     currency: ""
-  })
+  });
+  const [loading, setLoading] = useState(false);
 
   const [orderData, setOrderData] = useState(null);
+  const [orderDatesData, setOrderDatesData] = useState(null);
 
   const getOrderDetail = async () => {
+    setLoading(true);
+    const orderDates = await getOrderDates(orderId)
     const data = await getCheckoutOrderById(orderId);
+    setOrderDatesData(orderDates);
     setOrderData(data);
-  }
+    setLoading(false);
+  };
 
   useEffect(() => {
     if (orderData) {
       const currency_minor_unit = orderData?.totals?.currency_minor_unit;
+      const total_items = orderData?.totals?.total_items;
+      const total_tax = orderData?.totals?.total_tax;
       const cartSubTotal = getCorrectPrice(
-        parseInt(orderData?.totals?.total_items) + parseInt(orderData?.totals?.total_tax),
+        parseInt(total_items) + parseInt(total_tax),
         currency_minor_unit
       );
       const cartTotalDiscount = getCorrectPrice(
@@ -44,21 +53,19 @@ const Order = ({ orderId }) => {
         cartTotalDiscount,
         shippingTotal,
         currency: orderData?.totals?.currency_code
-      })
+      });
     }
-  }, [orderData])
+  }, [orderData]);
 
   useEffect(() => {
     getOrderDetail();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [orderId])
+  }, [orderId]);
 
   const getCorrectPrice = (number, currency_minor_unit) => {
     if (currency_minor_unit)
       return parseFloat((number / 100).toFixed(currency_minor_unit));
-
     return number;
-  }
+  };
 
   const {
     cartSubTotal,
@@ -67,94 +74,111 @@ const Order = ({ orderId }) => {
     currency
   } = cartData;
 
+  const { delivery_dates: deliveryDates, order_date: orderDate } = orderDatesData || {};
+
   return (
     <>
-      {orderData && (
-        <div className={styles.orderpage}>
-          <Header />
-          <section className={styles.ordersection}>
-            <h2 className={styles.orderhedaing}>
-              {oct.pageHeadingOrder}
-            </h2>
-            <div className={styles.ordercontainer}>
-              <div className={styles.orderCard}>
-                <div className={styles.orderDetail}>
-                  <p className={styles.label}>{oct.orderNo}</p>
-                  <p className={styles.value}>{orderData.order_data.number}</p>
+      {loading ? <Orderskeleton /> : (
+        <>
+          {orderData && (
+            <div className={styles.orderpage}>
+              <Header />
+              <section className={styles.ordersection}>
+                <h2 className={styles.orderhedaing}>
+                  {oct.pageHeadingOrder}
+                </h2>
+                <div className={styles.ordercontainer}>
+                  <div className={styles.orderCard}>
+                    <div className={styles.orderDetail}>
+                      <p className={styles.label}>{oct.orderNo}</p>
+                      <p className={styles.value}>{orderData.id}</p>
+                    </div>
+                    <div className={styles.orderDetail}>
+                      <p className={styles.label}>{oct.given}</p>
+                      <p className={styles.value}>{orderData.status}</p>
+                    </div>
+                    <div className={styles.orderDetail}>
+                      <p className={styles.label}>{oct.email}</p>
+                      <p className={styles.value}>{orderData.billing_address.email}</p>
+                    </div>
+                    <div className={styles.orderDetail}>
+                      <p className={styles.label}>{oct.total}</p>
+                      <p className={styles.value}>
+                        {currency} {getCorrectPrice(
+                          parseInt(orderData?.totals?.total_price),
+                          orderData?.totals?.currency_minor_unit
+                        )}
+                      </p>
+                    </div>
+                  </div>
+                  <aside className={styles.rightContainer}>
+                    <h4>{oct.orderInfromation}</h4>
+                    <table className={styles.table}>
+                      <thead>
+                        <tr>
+                          <th className={styles.firstHeading}>Item</th>
+                          <th>{oct.total}</th>
+                        </tr>
+                      </thead>
+                      <tbody>
+                        {orderData.items.map((item, index) => (
+                          <tr key={index}>
+                            <td>{item.name} <span>x {item.quantity}</span></td>
+                            <td>{currency} {getCorrectPrice(
+                              parseInt(item?.totals?.line_total) + parseInt(item?.totals?.line_total_tax),
+                              item?.totals?.currency_minor_unit
+                            )}</td>
+                          </tr>
+                        ))}
+                        <tr>
+                          <td>{oct.subtotal}</td>
+                          <td>{currency} {cartSubTotal}</td>
+                        </tr>
+                        <tr>
+                          <td>{oct.shipment}</td>
+                          <td>{currency} {shippingTotal}</td>
+                        </tr>
+                        <tr>
+                          <td>{oct.expectedDelivery}</td>
+                          <td>{deliveryDates && Object.keys(deliveryDates)[0]}</td>
+                        </tr>
+                        <tr>
+                          <td>{oct.orderDate}</td>
+                          <td>{orderDate}</td>
+                        </tr>
+                        {cartTotalDiscount && cartTotalDiscount > 0 ? (
+                          <tr>
+                            <td>{oct.rabat}</td>
+                            <td>{currency} {cartTotalDiscount}</td>
+                          </tr>
+                        ) : null}
+                      </tbody>
+                    </table>
+                  </aside>
                 </div>
-                <div className={styles.orderDetail}>
-                  <p className={styles.label}>{oct.given}</p>
-                  <p className={styles.value}>{new Date(orderData.order_data.date_created.date).toLocaleDateString()}</p>
+                <div className={styles.addresscontainer}>
+                  <div>
+                    <h4>{oct.billingAddress}</h4>
+                    <div>
+                      {orderData.billing_address && <AddressInfo address={orderData.billing_address} />}
+                    </div>
+                  </div>
+                  <div>
+                    <h4>{oct.deliveryAddress}</h4>
+                    <div>
+                      {orderData.shipping_address && <AddressInfo address={orderData.shipping_address} />}
+                    </div>
+                  </div>
                 </div>
-                <div className={styles.orderDetail}>
-                  <p className={styles.label}>{oct.email}</p>
-                  <p className={styles.value}>{orderData.order_data.billing.email}</p>
+                <div className={styles.continueBtnWrapper}>
+                  <Link href="/shop" className={styles.continueBtn} type="submit">
+                    {oct.continueShopping}
+                  </Link>
                 </div>
-                <div className={styles.orderDetail}>
-                  <p className={styles.label}>{oct.total}</p>
-                  <p className={styles.value} dangerouslySetInnerHTML={{ __html: orderData.order_data.total }} />
-                </div>
-              </div>
-              <aside className={styles.rightContainer}>
-                <h4>{oct.orderInfromation}</h4>
-                <table className={styles.table}>
-                  <thead>
-                    <tr>
-                      <th className={styles.firstHeading}>Item</th>
-                      <th>{oct.total}</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    {orderData.items.map((item, index) => (
-                      <tr key={index}>
-                        <td>{item.name} <span>x {item.quantity}</span></td>
-                        <td>{currency} {getCorrectPrice(
-                          parseInt(item?.totals?.line_total) + parseInt(item?.totals?.line_total_tax),
-                          item?.totals?.currency_minor_unit
-                        )}</td>
-                      </tr>
-                    ))}
-                    <tr>
-                      <td>{oct.subtotal}</td>
-                      <td>{currency} {cartSubTotal}</td>
-                    </tr>
-                    <tr>
-                      <td>{oct.shipment}</td>
-                      <td>{currency} {shippingTotal}</td>
-                    </tr>
-                    <tr>
-                      <td>{oct.expectedDelivery}</td>
-                      <td>{orderData.order_data.meta_data.find(meta => meta.key === 'delivery_date').value}</td>
-                    </tr>
-                    <tr>
-                      <td>{oct.rabat}</td>
-                      <td>{currency} {cartTotalDiscount}</td>
-                    </tr>
-                  </tbody>
-                </table>
-              </aside>
+              </section>
             </div>
-            <div className={styles.addresscontainer}>
-              <div>
-                <h4>{oct.billingAddress}</h4>
-                <div>
-                  {orderData.order_data.billing && <AddressInfo address={orderData.order_data.billing} />}
-                </div>
-              </div>
-              <div>
-                <h4>{oct.deliveryAddress}</h4>
-                <div>
-                  {orderData.order_data.shipping && <AddressInfo address={orderData.order_data.shipping} />}
-                </div>
-              </div>
-            </div>
-            <div className={styles.continueBtnWrapper}>
-              <Link href="/shop" className={styles.continueBtn} type="submit">
-                {oct.continueShopping}
-              </Link>
-            </div>
-          </section>
-        </div>
+          )}
+        </>
       )}
     </>
   );
