@@ -18,6 +18,9 @@ import Loader from "@/components/atoms/loader/loader";
 import { cartTranslation } from "@/locales";
 import { setShippingMethod } from "@/components/service/shipping";
 import CartShipping from "@/components/atoms/cartshiiping";
+import { applyLoader } from "@/helper/loader";
+import OverLayLoader from '@/components/atoms/overLayLoader';
+
 const lang = process.env.NEXT_PUBLIC_LANG || "dk";
 const toastTimer = parseInt(process.env.NEXT_PUBLIC_TOAST_TIMER);
 
@@ -39,6 +42,7 @@ function Cart() {
   const [loading, setLoading] = useState(false);
   const [progress, setProgress] = useState(0);
   const ct = cartTranslation[lang];
+  const [olLoader, setOlLoader] = useState(false);
 
   const [cartData, setCartData] = useState(INTIAL_CART_DATA);
 
@@ -209,7 +213,8 @@ function Cart() {
 
   };
 
-  const debouncedUpdateQuantity = quantityDebounce(updateQuantity, 1000);
+  // const debouncedUpdateQuantity = quantityDebounce(updateQuantity, 1000);
+  const debouncedUpdateQuantity = updateQuantity
 
   const checkShippingRates = useCallback(() => {
     return shipping && shipping.length > 0 && shipping[0]?.shipping?.[0]?.shipping_rates?.length > 0;
@@ -222,6 +227,7 @@ function Cart() {
           <Loader progress={progress} />
         </>
         : null}
+      {olLoader && <OverLayLoader />}
       {isCartReady ? (
         <div className={styles.shopingcart}>
           <Header />
@@ -240,12 +246,14 @@ function Cart() {
                           key: item_key,
                           images,
                           id: productId,
-                          prices,
+                          prices
                         } = cartItem;
+                        const itemTotals = cartItem.totals;
                         const currentImage =
                           images && images.length > 0 ? images[0].src : "";
                         const quantityValue = quantity;
                         const subtotal = getCorrectPrice(prices?.price);
+                        const itemsSubTotal = getCorrectPrice(parseInt(itemTotals?.line_subtotal) + parseInt(itemTotals?.line_subtotal_tax))
                         const subscription_schemes =
                           cartItem?.extensions?.subscription_schemes;
                         const current_options =
@@ -265,7 +273,11 @@ function Cart() {
                               <span
                                 className={styles.closebtn}
                                 onClick={(e) => {
-                                  removeProductCart(e, item_key);
+                                  applyLoader(
+                                    setOlLoader,
+                                    removeProductCart,
+                                    [e, item_key]
+                                  )
                                 }}
                               >
                                 <i className="fa-solid fa-xmark"></i>
@@ -304,10 +316,14 @@ function Cart() {
                                   <button
                                     className={`${styles.valueButton} ${styles.decreaseButton}`}
                                     onClick={() =>
-                                      debouncedUpdateQuantity(
-                                        cartItem.key,
-                                        -1,
-                                        cartItem.quantity
+                                      applyLoader(
+                                        setOlLoader,
+                                        debouncedUpdateQuantity,
+                                        [
+                                          cartItem.key,
+                                          -1,
+                                          cartItem.quantity
+                                        ]
                                       )
                                     }
                                   >
@@ -326,10 +342,14 @@ function Cart() {
                                   <button
                                     className={`${styles.valueButton} ${styles.increaseButton}`}
                                     onClick={() =>
-                                      debouncedUpdateQuantity(
-                                        cartItem.key,
-                                        1,
-                                        cartItem.quantity
+                                      applyLoader(
+                                        setOlLoader,
+                                        debouncedUpdateQuantity,
+                                        [
+                                          cartItem.key,
+                                          1,
+                                          cartItem.quantity
+                                        ]
                                       )
                                     }
                                   >
@@ -347,7 +367,13 @@ function Cart() {
                                   id="selectDelivery"
                                   value={selectedValue?.id}
                                   onChange={(e) =>
-                                    updateProductFrequency(e, cartItem.key)
+                                    applyLoader(
+                                      setOlLoader,
+                                      updateProductFrequency,
+                                      [
+                                        e, cartItem.key
+                                      ]
+                                    )
                                   }
                                   className={styles.selectbox}
                                 >
@@ -366,7 +392,7 @@ function Cart() {
                             <div className={styles.subtotalLeft}>
                               <p>{ct.subTotal}</p>
                               <p>
-                                {currency_symbol} {subtotal}
+                                {currency_symbol} {itemsSubTotal}
                               </p>
                             </div>
                           </Fragment>
@@ -397,7 +423,13 @@ function Cart() {
                       <button
                         type="submit"
                         className={styles.applydiscount}
-                        onClick={applyCoupon}
+                        onClick={() => {
+                          applyLoader(
+                            setOlLoader,
+                            applyCoupon,
+                            []
+                          )
+                        }}
                       >
                         {ct.applyDisscount}
                       </button>
@@ -424,7 +456,11 @@ function Cart() {
                             </label>
                             <span
                               onClick={() => {
-                                removeCoupon(coupons?.toUpperCase());
+                                applyLoader(
+                                  setOlLoader,
+                                  removeCoupon,
+                                  [coupons?.toUpperCase()]
+                                )
                               }}
                             >
                               X
@@ -468,7 +504,16 @@ function Cart() {
                         <CartShipping
                           shipping={shipping}
                           subscriptionShipping={subscriptionShipping}
-                          setCartShipment={setCartShipment}
+                          setCartShipment={
+                            (shipmentOpt, packageId) => {
+                              if (shipmentOpt && packageId)
+                                applyLoader(
+                                  setOlLoader,
+                                  setCartShipment,
+                                  [shipmentOpt, packageId]
+                                )
+                            }
+                          }
                           styles={styles}
                           getCorrectPrice={getCorrectPrice}
                         />
@@ -522,10 +567,11 @@ function Cart() {
               </p>
             </div>
           )}
-        </div>
+        </div >
       ) : (
         <Cartskeleton />
-      )}
+      )
+      }
     </>
   );
 }
