@@ -12,8 +12,6 @@ import Award from '@/components/atoms/Award/Awardlogo';
 import { addToCart, updateCartQuantity, removeCartItem } from '@/components/service/cart';
 import LazyLoad from "react-lazyload"; // Import LazyLoad component
 import CustomTestimonial from '@/components/atoms/CustomTestimonial/CustomTestimonial'
-import Categories from '@/components/atoms/categories';
-
 
 const HomePage = ({ homePageProps }) => {
   const { farmProducts, sessionalProducts, shippingMethods, vipPages, cutOffDaysDetail } = homePageProps;
@@ -28,17 +26,15 @@ const HomePage = ({ homePageProps }) => {
       {pathname === '/' && <TopImage farmProductProps={{ productData: farmProducts, enableMockData, addToCart, updateCartQuantity, removeCartItem }} />}
 
       <LazyLoad height={200} offset={100}>
-      <OrderDeadline {...shippingProps} />
-      {/* <Goods sessionalProductProps={{ productData: sessionalProducts, enableMockData, addToCart, updateCartQuantity, removeCartItem }} /> */}
-      {/* <ProductSlider /> */}
-      <Categories/>
-      <Register />
-      <AddSignUp {...vipProps} />
-      <VideoComponent />
-      <Award />
-     
-      {/* <Testimonials /> */}
-      <CustomTestimonial />
+        <OrderDeadline {...shippingProps} />
+        <Goods sessionalProductProps={{ productData: sessionalProducts, enableMockData, addToCart, updateCartQuantity, removeCartItem }} />
+        <ProductSlider />
+        <AddSignUp {...vipProps} />
+        <VideoComponent />
+        <Award />
+        <Register />
+        {/* <Testimonials /> */}
+        <CustomTestimonial />
 
       </LazyLoad>
     </>
