@@ -2,10 +2,14 @@ export const faqTranslation = {
     en: {
         fq: "FAQ",
         contact: "Contact us",
-        info: "info@fresh.land",
-        num: "53 79 07 07",
-        indikaj: "Indiakaj 20, 2100 Copenhagen Ø 🇩🇰",
+        info: "hej@fresh.land",
+        num: "+46 108 109 689",
+        indikaj: "Indiakaj 20, 2100 Köpenhamn Ø 🇩🇰",
         faqdelivery: "Delivery",
+        faqOrdering: "Ordering",
+        faqProduct: "Product",
+        faqOrganic: "Organic",
+        faqcustServ: "Customer service and contact",
     },
     dk: {
         fq: "dk FAQ",
