@@ -3,7 +3,7 @@ export const homepageTranslation = {
         freshlyHarvested: "Freshly Harvested",
         directlyFromFarm: "Directly from the Farm",
         lorem: "Freshly Harvested",
-        cardHeading: "Popular Goods",
+        cardHeading: "Directly from the Farm",
         freight: "Freight",
         weDeliverTo: "We deliver to Jutland and the fixed island as well Bornholm",
         seasonalGoods: "Seasonal Goods",
@@ -24,8 +24,7 @@ export const homepageTranslation = {
         unitedNations: "United Nations Good For All Winner (2021)",
         subscribeNewsletter: "Subscribe to newsletter",
         getNotified: "Get notified when your favorite product is in season",
-        email: "Email",
-        name:"Name",
+        email: "Email ID",
         register: "Register",
         iAgree: "I agree to the terms and conditions",
         contactUs: "Contact Us",
@@ -87,7 +86,7 @@ export const homepageTranslation = {
         unitedNations: "United Nations Good For All Winner (2021)",
         subscribeNewsletter: "Subscribe to newsletter",
         getNotified: "Get notified when your favorite product is in season",
-        email: "Email ",
+        email: "Email ID",
         register: "Register",
         iAgree: "I agree to the terms and conditions",
         contactUs: "Contact Us",
@@ -124,7 +123,7 @@ export const homepageTranslation = {
         unitedNations: "United Nations Good For All Winner (2021)",
         subscribeNewsletter: "Subscribe to newsletter",
         getNotified: "Get notified when your favorite product is in season",
-        email: "Email ",
+        email: "Email ID",
         register: "Register",
         iAgree: "I agree to the terms and conditions",
         contactUs: "Contact Us",
@@ -161,7 +160,7 @@ export const homepageTranslation = {
         unitedNations: "United Nations Good For All Winner (2021)",
         subscribeNewsletter: "Subscribe to newsletter",
         getNotified: "Get notified when your favorite product is in season",
-        email: "Email",
+        email: "Email ID",
         register: "Register",
         iAgree: "I agree to the terms and conditions",
         contactUs: "Contact Us",
