@@ -1,11 +1,11 @@
 import { KlarnaSuccessApiResponse } from "@/interfaces/klarnaOrderDetail";
-import axios from "axios";
+import axios from '@/utils/axios';
 
 const API_BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL;
-const CART_API_BASE_URL = `${API_BASE_URL}/wp-json/wc/store/v1/cart`;
-const INTERNAL_API_BASE_URL = `${API_BASE_URL}/wp-json/fl-cart/v1/cart`;
-const COUPON_CART_API_BASE_URL = `${API_BASE_URL}/wp-json/cocart/v3/cart`;
-const CHECKOUT_API_BASE_URL = `${API_BASE_URL}/wp-json/wc/store/v1`;
+const CART_API_BASE_URL = `/wp-json/wc/store/v1/cart`;
+const INTERNAL_API_BASE_URL = `/wp-json/fl-cart/v1/cart`;
+const COUPON_CART_API_BASE_URL = `/wp-json/cocart/v3/cart`;
+const CHECKOUT_API_BASE_URL = `/wp-json/wc/store/v1`;
 
 const updateOrderKlarnaSession = async (order_id: string, session_id: string) => {
     try {
