@@ -1,7 +1,6 @@
-import axios from "axios";
+import axios from '@/utils/axios';
 
-const API_BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL;
-const PAYMENT_API_BASE_URL = `${API_BASE_URL}/wp-json/fl-cart/v1/cart`;
+const PAYMENT_API_BASE_URL = `/wp-json/fl-cart/v1/cart`;
 
 const getPaymentGateways = async () => {
     try {
