@@ -1,7 +1,7 @@
-import axios from "axios";
+import axios from '@/utils/axios';
 
-const API_BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL;
-const PRODUCT_API_BASE_URL = `${API_BASE_URL}/wp-json/wc/v1`;
+// const API_BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL;
+const PRODUCT_API_BASE_URL = `/wp-json/wc/v1`;
 
 const getProducts = async () => {
     try {
