@@ -1,9 +1,8 @@
-import axios from "axios";
+import axios from '@/utils/axios';
 
-const API_BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL;
-const SHIPPING_API_BASE_URL = `${API_BASE_URL}/wp-json/wc/v1`;
-const CART_API_BASE_URL = `${API_BASE_URL}/wp-json/fl-cart/v1/cart`;
-const WOO_CART_API_BASE_URL = `${API_BASE_URL}/wp-json/wc/store/v1/cart`;
+const SHIPPING_API_BASE_URL = `/wp-json/wc/v1`;
+const CART_API_BASE_URL = `/wp-json/fl-cart/v1/cart`;
+const WOO_CART_API_BASE_URL = `/wp-json/wc/store/v1/cart`;
 
 const getShippingMethods = async () => {
     try {
