@@ -1,7 +1,6 @@
-import axios from "axios";
+import axios from '@/utils/axios';
 
-const API_BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL;
-const SHIPPING_API_BASE_URL = `${API_BASE_URL}/wp-json/wc/v1`;
+const SHIPPING_API_BASE_URL = `/wp-json/wc/v1`;
 
 const getVipPages = async (per_page = 3) => {
     try {
